15. Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-int i, j, k, sum;
+unsigned long int i, j, k, sum;
 i = 1;
 j = 2;
 sum = 2;
